[irep2] more split
