+++ src/irep2/irep2.cpp
@@ -7,6 +7,7 @@
 #include <irep2/irep2_type.h>
 #include <irep2/irep2_expr.h>
 #include <irep2/irep2_utils.h>
+#include <irep2/irep2_template_utils.h>
 #include <util/migrate.h>
 #include <util/std_types.h>
 #include <util/message/format.h>
@@ -15,293 +16,9 @@
 template <typename T>
 class register_irep_methods;
 
-std::string indent_str(unsigned int indent)
+std::string indent_str_irep2(unsigned int indent)
 {
   return std::string(indent, ' ');
-}
-
-template <class T>
-std::string pretty_print_func(unsigned int indent, std::string ident, T obj)
-{
-  list_of_memberst memb = obj.tostring(indent + 2);
-
-  std::string indentstr = indent_str(indent);
-  std::string exprstr = std::move(ident);
-
-  for(list_of_memberst::const_iterator it = memb.begin(); it != memb.end();
-      it++)
-  {
-    exprstr += "\n" + indentstr + "* " + it->first + " : " + it->second;
-  }
-
-  return exprstr;
-}
-
-/*************************** Base type2t definitions **************************/
-
-static const char *type_names[] = {
-  "bool",
-  "empty",
-  "symbol",
-  "struct",
-  "union",
-  "code",
-  "array",
-  "pointer",
-  "unsignedbv",
-  "signedbv",
-  "fixedbv",
-  "floatbv",
-  "string",
-  "cpp_name"};
-// If this fires, you've added/removed a type id, and need to update the list
-// above (which is ordered according to the enum list)
-static_assert(
-  sizeof(type_names) == (type2t::end_type_id * sizeof(char *)),
-  "Missing type name");
-
-std::string get_type_id(const type2t &type)
-{
-  return std::string(type_names[type.type_id]);
-}
-
-type2t::type2t(type_ids id)
-  : std::enable_shared_from_this<type2t>(), type_id(id), crc_val(0)
-{
-}
-
-bool type2t::operator==(const type2t &ref) const
-{
-  return cmpchecked(ref);
-}
-
-bool type2t::operator!=(const type2t &ref) const
-{
-  return !cmpchecked(ref);
-}
-
-bool type2t::operator<(const type2t &ref) const
-{
-  int tmp = type2t::lt(ref);
-  if(tmp < 0)
-    return true;
-  else if(tmp > 0)
-    return false;
-  else
-    return (lt(ref) < 0);
-}
-
-int type2t::ltchecked(const type2t &ref) const
-{
-  int tmp = type2t::lt(ref);
-  if(tmp != 0)
-    return tmp;
-
-  return lt(ref);
-}
-
-bool type2t::cmpchecked(const type2t &ref) const
-{
-  if(type_id == ref.type_id)
-    return cmp(ref);
-
-  return false;
-}
-
-int type2t::lt(const type2t &ref) const
-{
-  if(type_id < ref.type_id)
-    return -1;
-  if(type_id > ref.type_id)
-    return 1;
-  return 0;
-}
-
-std::string type2t::pretty(unsigned int indent) const
-{
-  return pretty_print_func<const type2t &>(indent, type_names[type_id], *this);
-}
-
-void type2t::dump() const
-{
-  default_message msg;
-  msg.debug(pretty(0));
-}
-
-size_t type2t::crc() const
-{
-  return do_crc();
-}
-
-size_t type2t::do_crc() const
-{
-  boost::hash_combine(this->crc_val, (uint8_t)type_id);
-  return this->crc_val;
-}
-
-void type2t::hash(crypto_hash &hash) const
-{
-  static_assert(type2t::end_type_id < 256, "Type id overflow");
-  uint8_t tid = type_id;
-  hash.ingest(&tid, sizeof(tid));
-}
-
-unsigned int bool_type2t::get_width() const
-{
-  // For the purpose of the byte representating memory model
-  return 8;
-}
-
-unsigned int bv_data::get_width() const
-{
-  return width;
-}
-
-unsigned int array_type2t::get_width() const
-{
-  // Two edge cases: the array can have infinite size, or it can have a dynamic
-  // size that's determined by the solver.
-  if(size_is_infinite)
-    throw new inf_sized_array_excp();
-
-  if(array_size->expr_id != expr2t::constant_int_id)
-    throw new dyn_sized_array_excp(array_size);
-
-  // Otherwise, we can multiply the size of the subtype by the number of elements.
-  unsigned int sub_width = subtype->get_width();
-
-  const expr2t *elem_size = array_size.get();
-  const constant_int2t *const_elem_size =
-    dynamic_cast<const constant_int2t *>(elem_size);
-  assert(const_elem_size != nullptr);
-  unsigned long num_elems = const_elem_size->as_ulong();
-
-  return num_elems * sub_width;
-}
-
-unsigned int pointer_type2t::get_width() const
-{
-  return config.ansi_c.pointer_width;
-}
-
-unsigned int empty_type2t::get_width() const
-{
-  throw new symbolic_type_excp();
-}
-
-unsigned int symbol_type2t::get_width() const
-{
-  assert(0 && "Fetching width of symbol type - invalid operation");
-  abort();
-}
-
-unsigned int cpp_name_type2t::get_width() const
-{
-  assert(0 && "Fetching width of cpp_name type - invalid operation");
-  abort();
-}
-
-unsigned int struct_type2t::get_width() const
-{
-  // Iterate over members accumulating width.
-  std::vector<type2tc>::const_iterator it;
-  unsigned int width = 0;
-  for(it = members.begin(); it != members.end(); it++)
-    width += (*it)->get_width();
-
-  return width;
-}
-
-unsigned int union_type2t::get_width() const
-{
-  // Iterate over members accumulating width.
-  std::vector<type2tc>::const_iterator it;
-  unsigned int width = 0;
-  for(it = members.begin(); it != members.end(); it++)
-    width = std::max(width, (*it)->get_width());
-
-  return width;
-}
-
-unsigned int fixedbv_type2t::get_width() const
-{
-  return width;
-}
-
-unsigned int floatbv_type2t::get_width() const
-{
-  return fraction + exponent + 1;
-}
-
-unsigned int code_data::get_width() const
-{
-  throw new symbolic_type_excp();
-}
-
-unsigned int string_type2t::get_width() const
-{
-  return width * 8;
-}
-
-unsigned int string_type2t::get_length() const
-{
-  return width;
-}
-
-const std::vector<type2tc> &struct_union_data::get_structure_members() const
-{
-  return members;
-}
-
-const std::vector<irep_idt> &
-struct_union_data::get_structure_member_names() const
-{
-  return member_names;
-}
-
-const irep_idt &struct_union_data::get_structure_name() const
-{
-  return name;
-}
-
-unsigned int struct_union_data::get_component_number(const irep_idt &comp) const
-{
-  unsigned int i = 0, count = 0, pos = 0;
-  for(auto const &it : member_names)
-  {
-    if(it == comp)
-    {
-      pos = i;
-      ++count;
-    }
-    i++;
-  }
-
-  if(count == 1)
-    return pos;
-
-  if(!count)
-  {
-    assert(
-      0 &&
-      fmt::format(
-        "Looking up index of nonexistant member \"{}\" in struct/union \"{}\"",
-        comp,
-        name)
-        .c_str());
-  }
-  else if(count > 1)
-  {
-    assert(
-      0 &&
-      fmt::format(
-        "Name \"{}\" matches more than one member\" in struct/union \"{}\"",
-        comp,
-        name)
-        .c_str());
-  }
-
-  abort();
 }
 
 namespace esbmct
@@ -314,275 +31,6 @@ auto type2t_traits<Args...>::make_contained(typename Args::result_type... args)
   return irep_container<base2t>(new derived(args...));
 }
 } // namespace esbmct
-
-/*************************** Base expr2t definitions **************************/
-
-expr2t::expr2t(const type2tc &_type, expr_ids id)
-  : std::enable_shared_from_this<expr2t>(), expr_id(id), type(_type), crc_val(0)
-{
-}
-
-expr2t::expr2t(const expr2t &ref)
-  : std::enable_shared_from_this<expr2t>(),
-    expr_id(ref.expr_id),
-    type(ref.type),
-    crc_val(ref.crc_val)
-{
-}
-
-bool expr2t::operator==(const expr2t &ref) const
-{
-  if(!expr2t::cmp(ref))
-    return false;
-
-  return cmp(ref);
-}
-
-bool expr2t::operator!=(const expr2t &ref) const
-{
-  return !(*this == ref);
-}
-
-bool expr2t::operator<(const expr2t &ref) const
-{
-  int tmp = expr2t::lt(ref);
-  if(tmp < 0)
-    return true;
-  else if(tmp > 0)
-    return false;
-  else
-    return (lt(ref) < 0);
-}
-
-unsigned long expr2t::depth() const
-{
-  unsigned long num_nodes = 0;
-
-  for(unsigned int idx = 0; idx < get_num_sub_exprs(); idx++)
-  {
-    const expr2tc *e = get_sub_expr(idx);
-    if(is_nil_expr(*e))
-      continue;
-    unsigned long tmp = (*e)->depth();
-    num_nodes = std::max(num_nodes, tmp);
-  }
-
-  num_nodes++; // Count ourselves.
-  return num_nodes;
-}
-
-unsigned long expr2t::num_nodes() const
-{
-  unsigned long count = 0;
-
-  for(unsigned int idx = 0; idx < get_num_sub_exprs(); idx++)
-  {
-    const expr2tc *e = get_sub_expr(idx);
-    if(is_nil_expr(*e))
-      continue;
-    count += (*e)->num_nodes();
-  }
-
-  count++; // Count ourselves.
-  return count;
-}
-
-int expr2t::ltchecked(const expr2t &ref) const
-{
-  int tmp = expr2t::lt(ref);
-  if(tmp != 0)
-    return tmp;
-
-  return lt(ref);
-}
-
-bool expr2t::cmp(const expr2t &ref) const
-{
-  if(expr_id != ref.expr_id)
-    return false;
-
-  if(type != ref.type)
-    return false;
-
-  return true;
-}
-
-int expr2t::lt(const expr2t &ref) const
-{
-  if(expr_id < ref.expr_id)
-    return -1;
-  if(expr_id > ref.expr_id)
-    return 1;
-
-  return type->ltchecked(*ref.type.get());
-}
-
-size_t expr2t::crc() const
-{
-  return do_crc();
-}
-
-size_t expr2t::do_crc() const
-{
-  boost::hash_combine(this->crc_val, type->do_crc());
-  boost::hash_combine(this->crc_val, (uint8_t)expr_id);
-  return this->crc_val;
-}
-
-void expr2t::hash(crypto_hash &hash) const
-{
-  static_assert(expr2t::end_expr_id < 256, "Expr id overflow");
-  uint8_t eid = expr_id;
-  hash.ingest(&eid, sizeof(eid));
-  type->hash(hash);
-}
-
-static const char *expr_names[] = {
-  "constant_int",
-  "constant_fixedbv",
-  "constant_floatbv",
-  "constant_bool",
-  "constant_string",
-  "constant_struct",
-  "constant_union",
-  "constant_array",
-  "constant_array_of",
-  "symbol",
-  "typecast",
-  "bitcast",
-  "nearbyint",
-  "if",
-  "equality",
-  "notequal",
-  "lessthan",
-  "greaterthan",
-  "lessthanequal",
-  "greaterthanequal",
-  "not",
-  "and",
-  "or",
-  "xor",
-  "implies",
-  "bitand",
-  "bitor",
-  "bitxor",
-  "bitnand",
-  "bitnor",
-  "bitnxor",
-  "bitnot",
-  "lshr",
-  "neg",
-  "abs",
-  "add",
-  "sub",
-  "mul",
-  "div",
-  "ieee_add",
-  "ieee_sub",
-  "ieee_mul",
-  "ieee_div",
-  "ieee_fma",
-  "ieee_sqrt",
-  "popcount",
-  "bswap",
-  "modulus",
-  "shl",
-  "ashr",
-  "dynamic_object",
-  "same_object",
-  "pointer_offset",
-  "pointer_object",
-  "address_of",
-  "byte_extract",
-  "byte_update",
-  "with",
-  "member",
-  "index",
-  "isnan",
-  "overflow",
-  "overflow_cast",
-  "overflow_neg",
-  "unknown",
-  "invalid",
-  "NULL-object",
-  "dereference",
-  "valid_object",
-  "deallocated_obj",
-  "dynamic_size",
-  "sideeffect",
-  "code_block",
-  "code_assign",
-  "code_init",
-  "code_decl",
-  "code_dead",
-  "code_printf",
-  "code_expression",
-  "code_return",
-  "code_skip",
-  "code_free",
-  "code_goto",
-  "object_descriptor",
-  "code_function_call",
-  "code_comma_id",
-  "invalid_pointer",
-  "code_asm",
-  "cpp_del_array",
-  "cpp_delete",
-  "cpp_catch",
-  "cpp_throw",
-  "cpp_throw_decl",
-  "cpp_throw_decl_end",
-  "isinf",
-  "isnormal",
-  "isfinite",
-  "signbit",
-  "concat",
-  "extract",
-};
-// If this fires, you've added/removed an expr id, and need to update the list
-// above (which is ordered according to the enum list)
-static_assert(
-  sizeof(expr_names) == (expr2t::end_expr_id * sizeof(char *)),
-  "Missing expr name");
-
-std::string get_expr_id(const expr2t &expr)
-{
-  return std::string(expr_names[expr.expr_id]);
-}
-
-std::string expr2t::pretty(unsigned int indent) const
-{
-  std::string ret =
-    pretty_print_func<const expr2t &>(indent, expr_names[expr_id], *this);
-  // Dump the type on the end.
-  ret += std::string("\n") + indent_str(indent) +
-         "* type : " + type->pretty(indent + 2);
-  return ret;
-}
-
-void expr2t::dump() const
-{
-  default_message msg;
-  msg.debug(pretty(0));
-}
-
-// Map a base type to it's list of names
-template <typename T>
-class base_to_names;
-
-template <>
-class base_to_names<type2t>
-{
-public:
-  static constexpr const char **names = type_names;
-};
-
-template <>
-class base_to_names<expr2t>
-{
-public:
-  static constexpr const char **names = expr_names;
-};
 
 // Undoubtedly a better way of doing this...
 namespace esbmct
@@ -613,101 +61,6 @@ auto expr2t_traits_always_construct<Args...>::make_contained(
 }
 } // namespace esbmct
 
-/**************************** Expression constructors *************************/
-
-unsigned long constant_int2t::as_ulong() const
-{
-  // XXXjmorse - add assertion that we don't exceed machine word width?
-  assert(!value.is_negative());
-  return value.to_uint64();
-}
-
-long constant_int2t::as_long() const
-{
-  // XXXjmorse - add assertion that we don't exceed machine word width?
-  return value.to_int64();
-}
-
-bool constant_bool2t::is_true() const
-{
-  return value;
-}
-
-bool constant_bool2t::is_false() const
-{
-  return !value;
-}
-
-std::string symbol_data::get_symbol_name() const
-{
-  switch(rlevel)
-  {
-  case level0:
-    return thename.as_string();
-  case level1:
-    return thename.as_string() + "?" + i2string(level1_num) + "!" +
-           i2string(thread_num);
-  case level2:
-    return thename.as_string() + "?" + i2string(level1_num) + "!" +
-           i2string(thread_num) + "&" + i2string(node_num) + "#" +
-           i2string(level2_num);
-  case level1_global:
-    // Just return global name,
-    return thename.as_string();
-  case level2_global:
-    // Global name with l2 details
-    return thename.as_string() + "&" + i2string(node_num) + "#" +
-           i2string(level2_num);
-  default:
-    assert(0 && "Unrecognized renaming level enum");
-    abort();
-  }
-}
-
-expr2tc constant_string2t::to_array() const
-{
-  std::vector<expr2tc> contents;
-  unsigned int length = value.as_string().size(), i;
-
-  type2tc type = get_uint8_type();
-
-  for(i = 0; i < length; i++)
-  {
-    constant_int2t *v = new constant_int2t(type, BigInt(value.as_string()[i]));
-    expr2tc ptr(v);
-    contents.push_back(ptr);
-  }
-
-  // Null terminator is implied.
-  contents.push_back(expr2tc(new constant_int2t(type, BigInt(0))));
-
-  unsignedbv_type2t *len_type = new unsignedbv_type2t(config.ansi_c.int_width);
-  type2tc len_tp(len_type);
-  constant_int2t *len_val = new constant_int2t(len_tp, BigInt(contents.size()));
-  expr2tc len_val_ref(len_val);
-
-  array_type2t *arr_type = new array_type2t(type, len_val_ref, false);
-  type2tc arr_tp(arr_type);
-  constant_array2t *a = new constant_array2t(arr_tp, contents);
-
-  expr2tc final_val(a);
-  return final_val;
-}
-
-const expr2tc &object_descriptor2t::get_root_object() const
-{
-  const expr2tc *tmp = &object;
-
-  do
-  {
-    if(is_member2t(*tmp))
-      tmp = &to_member2t(*tmp).source_value;
-    else if(is_index2t(*tmp))
-      tmp = &to_index2t(*tmp).source_value;
-    else
-      return *tmp;
-  } while(1);
-}
 
 // For CRCing to actually be accurate, expr/type ids mustn't overflow out of
 // a byte. If this happens then a) there are too many exprs, and b) the expr
@@ -715,628 +68,6 @@ const expr2tc &object_descriptor2t::get_root_object() const
 static_assert(type2t::end_type_id <= 256, "Type id overflow");
 static_assert(expr2t::end_expr_id <= 256, "Expr id overflow");
 
-static inline std::string type_to_string(const bool &thebool, int)
-{
-  return (thebool) ? "true" : "false";
-}
-
-static inline std::string
-type_to_string(const sideeffect_data::allockind &data, int)
-{
-  return (data == sideeffect_data::allockind::malloc)          ? "malloc"
-         : (data == sideeffect_data::allockind::realloc)       ? "realloc"
-         : (data == sideeffect_data::allockind::alloca)        ? "alloca"
-         : (data == sideeffect_data::allockind::cpp_new)       ? "cpp_new"
-         : (data == sideeffect_data::allockind::cpp_new_arr)   ? "cpp_new_arr"
-         : (data == sideeffect_data::allockind::nondet)        ? "nondet"
-         : (data == sideeffect_data::allockind::va_arg)        ? "va_arg"
-         : (data == sideeffect_data::allockind::function_call) ? "function_call"
-                                                               : "unknown";
-}
-
-static inline std::string type_to_string(const unsigned int &theval, int)
-{
-  char buffer[64];
-  snprintf(buffer, 63, "%d", theval);
-  return std::string(buffer);
-}
-
-static inline std::string
-type_to_string(const symbol_data::renaming_level &theval, int)
-{
-  switch(theval)
-  {
-  case symbol_data::level0:
-    return "Level 0";
-  case symbol_data::level1:
-    return "Level 1";
-  case symbol_data::level2:
-    return "Level 2";
-  case symbol_data::level1_global:
-    return "Level 1 (global)";
-  case symbol_data::level2_global:
-    return "Level 2 (global)";
-  default:
-    assert(0 && "Unrecognized renaming level enum");
-    abort();
-  }
-}
-
-static inline std::string type_to_string(const BigInt &theint, int)
-{
-  char buffer[256], *buf;
-
-  buf = theint.as_string(buffer, 256);
-  return std::string(buf);
-}
-
-static inline std::string type_to_string(const fixedbvt &theval, int)
-{
-  return theval.to_ansi_c_string();
-}
-
-static inline std::string type_to_string(const ieee_floatt &theval, int)
-{
-  return theval.to_ansi_c_string();
-}
-
-static inline std::string
-type_to_string(const std::vector<expr2tc> &theval, int indent)
-{
-  char buffer[64];
-  std::string astring = "\n";
-  int i;
-
-  i = 0;
-  for(auto const &it : theval)
-  {
-    snprintf(buffer, 63, "%d", i);
-    buffer[63] = '\0';
-    astring += indent_str(indent) + std::string(buffer) + ": " +
-               it->pretty(indent + 2) + "\n";
-    i++;
-  }
-
-  return astring;
-}
-
-static inline std::string
-type_to_string(const std::vector<type2tc> &theval, int indent)
-{
-  char buffer[64];
-  std::string astring = "\n";
-  int i;
-
-  i = 0;
-  for(auto const &it : theval)
-  {
-    snprintf(buffer, 63, "%d", i);
-    buffer[63] = '\0';
-    astring += indent_str(indent) + std::string(buffer) + ": " +
-               it->pretty(indent + 2) + "\n";
-    i++;
-  }
-
-  return astring;
-}
-
-static inline std::string
-type_to_string(const std::vector<irep_idt> &theval, int indent)
-{
-  char buffer[64];
-  std::string astring = "\n";
-  int i;
-
-  i = 0;
-  for(auto const &it : theval)
-  {
-    snprintf(buffer, 63, "%d", i);
-    buffer[63] = '\0';
-    astring +=
-      indent_str(indent) + std::string(buffer) + ": " + it.as_string() + "\n";
-    i++;
-  }
-
-  return astring;
-}
-
-static inline std::string type_to_string(const expr2tc &theval, int indent)
-{
-  if(theval.get() != nullptr)
-    return theval->pretty(indent + 2);
-  return "";
-}
-
-static inline std::string type_to_string(const type2tc &theval, int indent)
-{
-  if(theval.get() != nullptr)
-    return theval->pretty(indent + 2);
-  else
-    return "";
-}
-
-static inline std::string type_to_string(const irep_idt &theval, int)
-{
-  return theval.as_string();
-}
-
-static inline bool do_type_cmp(const bool &side1, const bool &side2)
-{
-  return (side1 == side2) ? true : false;
-}
-
-static inline bool
-do_type_cmp(const unsigned int &side1, const unsigned int &side2)
-{
-  return (side1 == side2) ? true : false;
-}
-
-static inline bool do_type_cmp(
-  const sideeffect_data::allockind &side1,
-  const sideeffect_data::allockind &side2)
-{
-  return (side1 == side2) ? true : false;
-}
-
-static inline bool do_type_cmp(
-  const symbol_data::renaming_level &side1,
-  const symbol_data::renaming_level &side2)
-{
-  return (side1 == side2) ? true : false;
-}
-
-static inline bool do_type_cmp(const BigInt &side1, const BigInt &side2)
-{
-  // BigInt has its own equality operator.
-  return (side1 == side2) ? true : false;
-}
-
-static inline bool do_type_cmp(const fixedbvt &side1, const fixedbvt &side2)
-{
-  return (side1 == side2) ? true : false;
-}
-
-static inline bool
-do_type_cmp(const ieee_floatt &side1, const ieee_floatt &side2)
-{
-  return (side1 == side2) ? true : false;
-}
-
-static inline bool do_type_cmp(
-  const std::vector<expr2tc> &side1,
-  const std::vector<expr2tc> &side2)
-{
-  return (side1 == side2);
-}
-
-static inline bool do_type_cmp(
-  const std::vector<type2tc> &side1,
-  const std::vector<type2tc> &side2)
-{
-  return (side1 == side2);
-}
-
-static inline bool do_type_cmp(
-  const std::vector<irep_idt> &side1,
-  const std::vector<irep_idt> &side2)
-{
-  return (side1 == side2);
-}
-
-static inline bool do_type_cmp(const expr2tc &side1, const expr2tc &side2)
-{
-  if(side1.get() == side2.get())
-    return true; // Catch null
-  else if(side1.get() == nullptr || side2.get() == nullptr)
-    return false;
-  else
-    return (side1 == side2);
-}
-
-static inline bool do_type_cmp(const type2tc &side1, const type2tc &side2)
-{
-  if(side1.get() == side2.get())
-    return true; // both null ptr check
-  if(side1.get() == nullptr || side2.get() == nullptr)
-    return false; // One of them is null, the other isn't
-  return (side1 == side2);
-}
-
-static inline bool do_type_cmp(const irep_idt &side1, const irep_idt &side2)
-{
-  return (side1 == side2);
-}
-
-static inline bool
-do_type_cmp(const type2t::type_ids &, const type2t::type_ids &)
-{
-  return true; // Dummy field comparison.
-}
-
-static inline bool
-do_type_cmp(const expr2t::expr_ids &, const expr2t::expr_ids &)
-{
-  return true; // Dummy field comparison.
-}
-
-static inline int do_type_lt(const bool &side1, const bool &side2)
-{
-  if(side1 < side2)
-    return -1;
-  else if(side2 < side1)
-    return 1;
-  else
-    return 0;
-}
-
-static inline int
-do_type_lt(const unsigned int &side1, const unsigned int &side2)
-{
-  if(side1 < side2)
-    return -1;
-  else if(side2 < side1)
-    return 1;
-  else
-    return 0;
-}
-
-static inline int do_type_lt(
-  const sideeffect_data::allockind &side1,
-  const sideeffect_data::allockind &side2)
-{
-  if(side1 < side2)
-    return -1;
-  else if(side2 < side1)
-    return 1;
-  else
-    return 0;
-}
-
-static inline int do_type_lt(
-  const symbol_data::renaming_level &side1,
-  const symbol_data::renaming_level &side2)
-{
-  if(side1 < side2)
-    return -1;
-  else if(side2 < side1)
-    return 1;
-  else
-    return 0;
-}
-
-static inline int do_type_lt(const BigInt &side1, const BigInt &side2)
-{
-  // BigInt also has its own less than comparator.
-  return side1.compare(side2);
-}
-
-static inline int do_type_lt(const fixedbvt &side1, const fixedbvt &side2)
-{
-  if(side1 < side2)
-    return -1;
-  else if(side1 > side2)
-    return 1;
-  return 0;
-}
-
-static inline int do_type_lt(const ieee_floatt &side1, const ieee_floatt &side2)
-{
-  if(side1 < side2)
-    return -1;
-  else if(side1 > side2)
-    return 1;
-  return 0;
-}
-
-static inline int
-do_type_lt(const std::vector<expr2tc> &side1, const std::vector<expr2tc> &side2)
-{
-  int tmp = 0;
-  std::vector<expr2tc>::const_iterator it2 = side2.begin();
-  for(auto const &it : side1)
-  {
-    tmp = it->ltchecked(**it2);
-    if(tmp != 0)
-      return tmp;
-    it2++;
-  }
-  return 0;
-}
-
-static inline int
-do_type_lt(const std::vector<type2tc> &side1, const std::vector<type2tc> &side2)
-{
-  if(side1.size() < side2.size())
-    return -1;
-  else if(side1.size() > side2.size())
-    return 1;
-
-  int tmp = 0;
-  std::vector<type2tc>::const_iterator it2 = side2.begin();
-  for(auto const &it : side1)
-  {
-    tmp = it->ltchecked(**it2);
-    if(tmp != 0)
-      return tmp;
-    it2++;
-  }
-  return 0;
-}
-
-static inline int do_type_lt(
-  const std::vector<irep_idt> &side1,
-  const std::vector<irep_idt> &side2)
-{
-  if(side1 < side2)
-    return -1;
-  else if(side2 < side1)
-    return 1;
-  return 0;
-}
-
-static inline int do_type_lt(const expr2tc &side1, const expr2tc &side2)
-{
-  if(side1.get() == side2.get())
-    return 0; // Catch nulls
-  else if(side1.get() == nullptr)
-    return -1;
-  else if(side2.get() == nullptr)
-    return 1;
-  else
-    return side1->ltchecked(*side2.get());
-}
-
-static inline int do_type_lt(const type2tc &side1, const type2tc &side2)
-{
-  if(*side1.get() == *side2.get())
-    return 0; // Both may be null;
-  else if(side1.get() == nullptr)
-    return -1;
-  else if(side2.get() == nullptr)
-    return 1;
-  else
-    return side1->ltchecked(*side2.get());
-}
-
-static inline int do_type_lt(const irep_idt &side1, const irep_idt &side2)
-{
-  if(side1 < side2)
-    return -1;
-  if(side2 < side1)
-    return 1;
-  return 0;
-}
-
-static inline int do_type_lt(const type2t::type_ids &, const type2t::type_ids &)
-{
-  return 0; // Dummy field comparison
-}
-
-static inline int do_type_lt(const expr2t::expr_ids &, const expr2t::expr_ids &)
-{
-  return 0; // Dummy field comparison
-}
-
-static inline size_t do_type_crc(const bool &theval)
-{
-  return boost::hash<bool>()(theval);
-}
-
-static inline void do_type_hash(const bool &thebool, crypto_hash &hash)
-{
-  if(thebool)
-  {
-    uint8_t tval = 1;
-    hash.ingest(&tval, sizeof(tval));
-  }
-  else
-  {
-    uint8_t tval = 0;
-    hash.ingest(&tval, sizeof(tval));
-  }
-}
-
-static inline size_t do_type_crc(const unsigned int &theval)
-{
-  return boost::hash<unsigned int>()(theval);
-}
-
-static inline void do_type_hash(const unsigned int &theval, crypto_hash &hash)
-{
-  hash.ingest((void *)&theval, sizeof(theval));
-}
-
-static inline size_t do_type_crc(const sideeffect_data::allockind &theval)
-{
-  return boost::hash<uint8_t>()(theval);
-}
-
-static inline void
-do_type_hash(const sideeffect_data::allockind &theval, crypto_hash &hash)
-{
-  hash.ingest((void *)&theval, sizeof(theval));
-}
-
-static inline size_t do_type_crc(const symbol_data::renaming_level &theval)
-{
-  return boost::hash<uint8_t>()(theval);
-}
-
-static inline void
-do_type_hash(const symbol_data::renaming_level &theval, crypto_hash &hash)
-{
-  hash.ingest((void *)&theval, sizeof(theval));
-}
-
-static inline size_t do_type_crc(const BigInt &theint)
-{
-  if(theint.is_zero())
-    return boost::hash<uint8_t>()(0);
-
-  size_t crc = 0;
-  std::array<unsigned char, 256> buffer;
-  if(theint.dump(buffer.data(), buffer.size()))
-  {
-    for(unsigned int i = 0; i < buffer.size(); i++)
-      boost::hash_combine(crc, buffer[i]);
-  }
-  else
-  {
-    // bigint is too large to fit in that static buffer. This is insane; but
-    // rather than wasting time heap allocing we'll just skip recording data,
-    // at the price of possible crc collisions.
-    ;
-  }
-  return crc;
-}
-
-static inline void do_type_hash(const BigInt &theint, crypto_hash &hash)
-{
-  // Zero has no data in bigints.
-  if(theint.is_zero())
-  {
-    uint8_t val = 0;
-    hash.ingest(&val, sizeof(val));
-    return;
-  }
-
-  std::array<unsigned char, 256> buffer;
-  if(theint.dump(buffer.data(), buffer.size()))
-  {
-    hash.ingest(buffer.data(), buffer.size());
-  }
-  else
-  {
-    // bigint is too large to fit in that static buffer. This is insane; but
-    // rather than wasting time heap allocing we'll just skip recording data,
-    // at the price of possible crc collisions.
-    ;
-  }
-}
-
-static inline size_t do_type_crc(const fixedbvt &theval)
-{
-  return do_type_crc(BigInt(theval.to_ansi_c_string().c_str()));
-}
-
-static inline void do_type_hash(const fixedbvt &theval, crypto_hash &hash)
-{
-  do_type_hash(BigInt(theval.to_ansi_c_string().c_str()), hash);
-}
-
-static inline size_t do_type_crc(const ieee_floatt &theval)
-{
-  return do_type_crc(theval.pack());
-}
-
-static inline void do_type_hash(const ieee_floatt &theval, crypto_hash &hash)
-{
-  do_type_hash(theval.pack(), hash);
-}
-
-static inline size_t do_type_crc(const std::vector<expr2tc> &theval)
-{
-  size_t crc = 0;
-  for(auto const &it : theval)
-    boost::hash_combine(crc, it->do_crc());
-
-  return crc;
-}
-
-static inline void
-do_type_hash(const std::vector<expr2tc> &theval, crypto_hash &hash)
-{
-  for(auto const &it : theval)
-    it->hash(hash);
-}
-
-static inline size_t do_type_crc(const std::vector<type2tc> &theval)
-{
-  size_t crc = 0;
-  for(auto const &it : theval)
-    boost::hash_combine(crc, it->do_crc());
-
-  return crc;
-}
-
-static inline void
-do_type_hash(const std::vector<type2tc> &theval, crypto_hash &hash)
-{
-  for(auto const &it : theval)
-    it->hash(hash);
-}
-
-static inline size_t do_type_crc(const std::vector<irep_idt> &theval)
-{
-  size_t crc = 0;
-  for(auto const &it : theval)
-    boost::hash_combine(crc, it.as_string());
-
-  return crc;
-}
-
-static inline void
-do_type_hash(const std::vector<irep_idt> &theval, crypto_hash &hash)
-{
-  for(auto const &it : theval)
-    hash.ingest((void *)it.as_string().c_str(), it.as_string().size());
-}
-
-static inline size_t do_type_crc(const expr2tc &theval)
-{
-  if(theval.get() != nullptr)
-    return theval->do_crc();
-  return boost::hash<uint8_t>()(0);
-}
-
-static inline void do_type_hash(const expr2tc &theval, crypto_hash &hash)
-{
-  if(theval.get() != nullptr)
-    theval->hash(hash);
-}
-
-static inline size_t do_type_crc(const type2tc &theval)
-{
-  if(theval.get() != nullptr)
-    return theval->do_crc();
-  return boost::hash<uint8_t>()(0);
-}
-
-static inline void do_type_hash(const type2tc &theval, crypto_hash &hash)
-{
-  if(theval.get() != nullptr)
-    theval->hash(hash);
-}
-
-static inline size_t do_type_crc(const irep_idt &theval)
-{
-  return boost::hash<std::string>()(theval.as_string());
-}
-
-static inline void do_type_hash(const irep_idt &theval, crypto_hash &hash)
-{
-  hash.ingest((void *)theval.as_string().c_str(), theval.as_string().size());
-}
-
-static inline size_t do_type_crc(const type2t::type_ids &i)
-{
-  return boost::hash<uint8_t>()(i);
-}
-
-static inline void do_type_hash(const type2t::type_ids &, crypto_hash &)
-{
-  // Dummy field crc
-}
-
-static inline size_t do_type_crc(const expr2t::expr_ids &i)
-{
-  return boost::hash<uint8_t>()(i);
-}
-
-static inline void do_type_hash(const expr2t::expr_ids &, crypto_hash &)
-{
-  // Dummy field crc
-}
 
 template <typename T>
 void do_type2string(
