+++ src/irep2/irep2_type.h
@@ -4,6 +4,7 @@
 #include <irep2/irep2.h>
 #include <util/type.h>
 
+
 // Start with forward class definitions
 
 class bool_type2t;
