+++ src/irep2/irep2_utils.h
@@ -6,6 +6,32 @@
 #include <irep2/irep2_expr.h>
 #include <util/migrate.h>
 
+
+std::string indent_str_irep2(unsigned int indent);
+
+// Map a base type to it's list of names
+template <typename T>
+class base_to_names;
+
+
+template <class T>
+std::string pretty_print_func(unsigned int indent, std::string ident, T obj)
+{
+  list_of_memberst memb = obj.tostring(indent + 2);
+
+  std::string indentstr = indent_str_irep2(indent);
+  std::string exprstr = std::move(ident);
+
+  for(list_of_memberst::const_iterator it = memb.begin(); it != memb.end();
+      it++)
+  {
+    exprstr += "\n" + indentstr + "* " + it->first + " : " + it->second;
+  }
+
+  return exprstr;
+}
+
+
 /** Test whether type is an integer. */
 inline bool is_bv_type(const type2tc &t)
 {
